+++ BackupR/Verbs/CleanupVerb.cs
@@ -8,7 +8,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Tekook.BackupR.Lib.Config;
 using Tekook.BackupR.Lib.Contracts;
-using Tekook.BackupR.Lib.ProviderExtensions;
+using Tekook.BackupR.Lib.Extensions;
 using Tekook.BackupR.Lib.StateManagement;
 using Tekook.VerbR.Resolvers;
